Update random_number.py
Add functionality so that if a seed of -1 is provided, it will generate a new random number each time the chain is run.
This was originally a feature when the random number node was created, but it was removed at the time due to issues with caching in iterators. The derive seed node resolved the issues around iterators, so it makes sense to add this back in.