Support NCNN CPU inference
This is a useful option on systems with strong CPU and weak GPU, where
PyTorch is nontrivial to install (e.g. ppc64le systems).

Currently CPU inference is keyed on which NCNN package is
installed; a future PR could allow CPU inference even when the Vulkan
NCNN package is installed.