use native clipboard on macos (#2031)
* use native clipboard on macos

On macOS we need the native Edit menu to make all keyboard shortcuts accessable.

The shortcuts can copy/cut/paste text and nodes. But using entries from the menu
itself only interacts with text.

Therefore on macOS a new menu is added: "Nodes"

Menu structure:

*Nodes
|
|- Undo
|- Redo
|--------
|- Cut
|- Copy
|- Paste
|--------
|- Duplicate
|- Duplicate with Connections

This menus sole purpose is to give the user a menu to interact with Nodes.

closes #1590

* added a Window menu on mac moved View menu

The View Menu is now located between Edit and Nodes to align it more with the default menu structure on macOS.

Also the Window Menu used on Macs has been added.