Create "Premultiply Node"
This node premultiplies the RGB channels with the Alpha channel and produces an image with an "Associated pixel state". 

A lot of the graphics/workflows out there "assume" premultiplied RGBA channels so having this node available in chaiNNer's toolset will help it do more conversions that previously weren't possible in the application.