Feature - Add MiniMap to chaiNNer (#2596)
* Add Minimap

This adds a minimap to the chaiNNer schematic

* Update global.scss - adds Minimap BG Color

This tweaks the default miniMap BG Color to fit in with the default chaiNNer color scheme

* Update ReactFlowBox.tsx

lint fixes

* Update ReactFlowBox.tsx

more linting, now auto fixed

* Update global.scss

And now with even more linting! When will it ever stop?

* Update global.scss - Minimap

Tweak and move MiniMap background colors to the respective themes

* Update ReactFlowBox.tsx - MiniMap

Reduce zoomScroll strength inside of the MiniMap

* Update ReactFlowBox.tsx

Fix typo!

* Update ReactFlowBox.tsx - MiniMap

Move the basic MiniMap theming out of the ReactFlow setup and into the global.scss where we can define it for each application "Theme".

* Update global.scss - MiniMap

Update the MiniMap Theming and define them for each Application Theme.