Create "Premultiplied Alpha" Node (#2578)
* Create "Premultiply Node"

This node premultiplies the RGB channels with the Alpha channel and produces an image with an "Associated pixel state". 

A lot of the graphics/workflows out there "assume" premultiplied RGBA channels so having this node available in chaiNNer's toolset will help it do more conversions that previously weren't possible in the application.

* Update premultiply.py

linting

* Delete backend/src/packages/chaiNNer_standard/image_adjustment/adjustments/premultiply.py

* Create "Premultiplied Alpha" node

This node allows the user to convert from a "Premultiplied" Alpha channel state to a "Straight/Unpremultiplied" alpha channel state, and vice versa.

* Update premultiplied_alpha.py

Try and just force dtype to float32 to clear the backend linter.

* Update premultiplied_alpha.py - Reworked

Reworked the PR to use the suggested naming and concepts.